update ui and routing

+++ back-end/friends/views.py
@@ -15,15 +15,20 @@ class friends_viewset(viewsets.ModelViewSet):
     serializer_class = friendListSerializer
 
     def list(self, request):
-        queryset = friendList.objects.filter(
-            ((Q(user1=request.user) & Q(user1_blocked_user2=False)) |
-            (Q(user2=request.user) & Q(user2_blocked_user1=False))) &
-            Q(are_friends=True)
-        )
+        if request.user.is_authenticated:
+            queryset = friendList.objects.filter(
+                ((Q(user1=request.user) & Q(user1_blocked_user2=False)) |
+                (Q(user2=request.user) & Q(user2_blocked_user1=False))) &
+                Q(are_friends=True)
+            )
+        else:
+            return Response({"error": "User not authenticated"}, status=status.HTTP_404_NOT_FOUND)
         serializer = friendListSerializer(queryset, many=True)
         return Response(serializer.data)
 
     def create(self, request):
+        if not request.user.is_authenticated:
+            return Response({"error": "User not authenticated"}, status=status.HTTP_401_UNAUTHORIZED)
         serializer = friendListSerializer(data=request.data)
         if serializer.is_valid():
             user1 = serializer.validated_data['user1']
@@ -40,6 +45,8 @@ class friends_viewset(viewsets.ModelViewSet):
 
     def destroy(self, request, pk=None):
         try:
+            if not request.user.is_authenticated:
+                return Response({"error": "User not authenticated"}, status=status.HTTP_401_UNAUTHORIZED)
             queryset = friendList.objects.get(pk=pk)
             if queryset is not None and queryset.are_friends == True:
                 queryset.delete()
@@ -51,6 +58,8 @@ class friends_viewset(viewsets.ModelViewSet):
     @action(detail=False, methods=['GET', 'POST'])
     def block(self, request, pk=None):
         if request.method == 'POST':
+            if not request.user.is_authenticated:
+                return Response({"error": "User not authenticated"}, status=status.HTTP_401_UNAUTHORIZED)
             serializer = friendListSerializer(data=request.data)
             if serializer.is_valid():
                 user1 = serializer.validated_data['user1']
@@ -65,10 +74,13 @@ class friends_viewset(viewsets.ModelViewSet):
                     friendShip.save()
                     return Response(status=200)
         if request.method == 'GET':
-            queryset = friendList.objects.filter(
-                (Q(user1=request.user) & Q(user1_blocked_user2=True)) |
-                (Q(user2=request.user) & Q(user2_blocked_user1=True))
-            )
+            if request.user.is_authenticated:
+                queryset = friendList.objects.filter(
+                    (Q(user1=request.user) & Q(user1_blocked_user2=True)) |
+                    (Q(user2=request.user) & Q(user2_blocked_user1=True))
+                )
+            else:
+                return Response({"error": "User not authenticated"}, status=status.HTTP_401_UNAUTHORIZED)
             blockList = friendListSerializer(queryset, many=True)
             return Response(blockList.data)
         return Response(status=400)
@@ -76,10 +88,13 @@ class friends_viewset(viewsets.ModelViewSet):
     @action(detail=False, methods=['GET', 'POST'])
     def unblock(self, request, pk=None):
         if request.method == 'POST':
-            queryset = friendList.objects.filter(
-            (Q(user1=request.user) & Q(user1_blocked_user2=True)) |
-            (Q(user2=request.user) & Q(user2_blocked_user1=True))
-        )
+            if request.user.is_authenticated:
+                queryset = friendList.objects.filter(
+                (Q(user1=request.user) & Q(user1_blocked_user2=True)) |
+                (Q(user2=request.user) & Q(user2_blocked_user1=True))
+            )
+            else:
+                return Response({"error": "User not authenticated"}, status=status.HTTP_401_UNAUTHORIZED)
             if queryset:
                 queryset.delete()
                 return Response(status=200)
@@ -88,10 +103,13 @@ class friends_viewset(viewsets.ModelViewSet):
     @action(detail=False, methods=['GET', 'POST'])
     def accept(self, request, pk=None):
         if request.method == 'POST':
-            queryset = friendList.objects.filter(
-                (Q(user1=request.user) & Q(user2_invited_user1=True)) |
-                (Q(user2=request.user) & Q(user1_invited_user2=True))
-            ).first()
+            if request.user.is_authenticated:
+                queryset = friendList.objects.filter(
+                    (Q(user1=request.user) & Q(user2_invited_user1=True)) |
+                    (Q(user2=request.user) & Q(user1_invited_user2=True))
+                ).first()
+            else:
+                return Response({"error": "User not authenticated"}, status=status.HTTP_401_UNAUTHORIZED)
             if not queryset:
                 return Response({"detail": "No pending invitation found."}, status=400)
             queryset.are_friends = True
@@ -101,10 +119,13 @@ class friends_viewset(viewsets.ModelViewSet):
             serializer = friendListSerializer(queryset)
             return Response(serializer.data, status=200)
         if request.method == 'GET':
-            queryset = friendList.objects.filter(
-                (Q(user1=request.user) & Q(user2_invited_user1=True)) |
-                (Q(user2=request.user) & Q(user1_invited_user2=True))
-            )
+            if request.user.is_authenticated:
+                queryset = friendList.objects.filter(
+                    (Q(user1=request.user) & Q(user2_invited_user1=True)) |
+                    (Q(user2=request.user) & Q(user1_invited_user2=True))
+                )
+            else:
+                return Response({"error": "User not authenticated"}, status=status.HTTP_401_UNAUTHORIZED)
             if queryset:
                 serializer = friendListSerializer(queryset, many=True)
                 return (Response(serializer.data))
